refactor(cli): extract remember + cli-hints; fix YAML-injection in frontmatter
`src/cli.ts` was 2,687 lines and carried memory-domain logic that
belonged in dedicated modules. This commit extracts:

- **`src/remember.ts`** (224 lines): `parseDuration`,
  `buildMemoryFrontmatter`, `runAutoHeuristics`, `runLlmEnrich`,
  `readMemoryContent`. The `remember` command in cli.ts now imports
  these as named exports — testable in isolation, no test required to
  spawn the CLI binary.
- **`src/cli-hints.ts`** (303 lines): the `EMBEDDED_HINTS` /
  `EMBEDDED_HINTS_FULL` template literals (~250 lines of agent-onboarding
  content that has nothing to do with argument parsing).
- **`src/common.ts`** gains `tryReadStdinText` (was a private helper in
  cli.ts; needed by both `readMemoryContent` and the unchanged
  `readKnowledgeContent`).

Result: cli.ts drops to 2,212 lines (-475).

## YAML-injection fix (security review High finding)

`buildMemoryFrontmatter` previously interpolated user-supplied strings
directly into YAML lines:

    description: ${fields.description}

A user passing a `description` containing `\n` + `tags: [pwned]` would
inject extra frontmatter keys that the indexer/agent would then trust.
The new implementation builds an object and runs it through
`yaml.stringify`, which safely quotes any value containing newlines or
YAML metacharacters. A regression test (`description containing
newlines + forged tags cannot inject extra keys`) locks the fix in.

## Tests

`tests/remember-unit.test.ts` — 19 cases covering `parseDuration`,
`buildMemoryFrontmatter` (including the YAML-injection guard, source
metacharacter round-trip, empty-fields case, whitespace-only fields),
and `runAutoHeuristics` (code/url/iso-date/today/empty/first-person
detection). Documents the existing case-sensitivity of the first-person
match so a future widening will land deliberately.

Implements two of the "top 3 before v1" picks from #177:
- "Extract `remember` + inline hints out of `cli.ts`" (architecture #1)
- "buildMemoryFrontmatter interpolates raw strings into YAML" (security H#3)

bun test → 1641 pass / 0 fail / 7 skip. tsc + biome clean.
