+++ src/cli.ts
@@ -3,7 +3,8 @@ import fs from "node:fs";
 import path from "node:path";
 import { defineCommand, runMain } from "citty";
 import { deriveCanonicalAssetName, resolveAssetPathFromName } from "./asset-spec";
-import { isWithin, resolveStashDir } from "./common";
+import { EMBEDDED_HINTS, EMBEDDED_HINTS_FULL } from "./cli-hints";
+import { isWithin, resolveStashDir, tryReadStdinText } from "./common";
 import { generateBashCompletions, installBashCompletions } from "./completions";
 import type { RegistryConfigEntry } from "./config";
 import { DEFAULT_CONFIG, getConfigPath, loadConfig, loadUserConfig, saveConfig } from "./config";
@@ -22,6 +23,7 @@ import { formatPlain, outputJsonl } from "./output-text";
 import { getCacheDir, getDbPath, getDefaultStashDir } from "./paths";
 import { buildRegistryIndex, writeRegistryIndex } from "./registry-build-index";
 import { searchRegistry } from "./registry-search";
+import { buildMemoryFrontmatter, parseDuration, readMemoryContent, runAutoHeuristics, runLlmEnrich } from "./remember";
 import { checkForUpdate, performUpgrade } from "./self-update";
 import { akmAdd } from "./stash-add";
 import { akmClone } from "./stash-clone";
@@ -849,12 +851,6 @@ const feedbackCommand = defineCommand({
   },
 });
 
-function tryReadStdinText(): string | undefined {
-  if (process.stdin.isTTY) return undefined;
-  const input = fs.readFileSync(0, "utf8");
-  return input.length > 0 ? input : undefined;
-}
-
 function normalizeMarkdownAssetName(name: string | undefined, fallback: string): string {
   const trimmed = (name ?? fallback)
     .trim()
@@ -886,14 +882,6 @@ function inferAssetName(content: string, fallbackPrefix: string, preferred?: str
     .find((line) => line.length > 0);
   const basis = preferred?.trim() || firstNonEmptyLine || fallbackPrefix;
   return slugifyAssetName(basis, fallbackPrefix);
-}
-
-function readMemoryContent(contentArg: string | undefined): string {
-  const content = contentArg ?? tryReadStdinText();
-  if (!content?.trim()) {
-    throw new UsageError("Memory content is required. Pass quoted text or pipe markdown into stdin.");
-  }
-  return content;
 }
 
 function readKnowledgeContent(source: string): { content: string; preferredName?: string } {
@@ -1164,177 +1152,6 @@ const workflowCommand = defineCommand({
     });
   },
 });
-
-// ── Memory frontmatter helpers ────────────────────────────────────────────────
-
-/**
- * Parse a shorthand duration string to a number of milliseconds.
- * Supports: `30d` (days), `12h` (hours), `6m` (months, approximated as 30d).
- */
-function parseDuration(s: string): number {
-  const match = s.trim().match(/^(\d+)([dhm])$/i);
-  if (!match) throw new UsageError(`Invalid --expires format "${s}". Use shorthand like 30d, 12h, or 6m.`);
-  const n = Number(match[1]);
-  const unit = match[2].toLowerCase();
-  if (unit === "d") return n * 24 * 60 * 60 * 1000;
-  if (unit === "h") return n * 60 * 60 * 1000;
-  // 'm' = months, approximated as 30 days
-  return n * 30 * 24 * 60 * 60 * 1000;
-}
-
-/**
- * Build a YAML frontmatter block from memory metadata.
- * Only includes fields that are present (non-empty).
- */
-function buildMemoryFrontmatter(fields: {
-  description?: string;
-  tags?: string[];
-  source?: string;
-  observed_at?: string;
-  expires?: string;
-  subjective?: boolean;
-}): string {
-  const lines: string[] = ["---"];
-  if (fields.description) lines.push(`description: ${fields.description}`);
-  if (fields.tags && fields.tags.length > 0) {
-    lines.push(`tags: [${fields.tags.join(", ")}]`);
-  }
-  if (fields.source) lines.push(`source: ${fields.source}`);
-  if (fields.observed_at) lines.push(`observed_at: ${fields.observed_at}`);
-  if (fields.expires) lines.push(`expires: ${fields.expires}`);
-  if (fields.subjective) lines.push("subjective: true");
-  lines.push("---");
-  return lines.join("\n");
-}
-
-/**
- * Run heuristic analysis on memory body text.
- * Returns derived metadata fields without modifying any files.
- */
-function runAutoHeuristics(body: string): {
-  tags: string[];
-  source?: string;
-  observed_at?: string;
-  subjective?: boolean;
-} {
-  const tags: string[] = [];
-
-  // Fenced code block present → tag "code"
-  if (/^```/m.test(body)) {
-    tags.push("code");
-  }
-
-  // First-person pronoun → subjective
-  const subjective = /\b(I|we|my|our)\b/.test(body) ? true : undefined;
-
-  // First URL-shaped token → source
-  const urlMatch = body.match(/https?:\/\/[^\s)>'"]+/);
-  const source = urlMatch ? urlMatch[0] : undefined;
-
-  // ISO date token or obvious relative date phrase → observed_at
-  let observed_at: string | undefined;
-  const isoMatch = body.match(/\b(\d{4}-\d{2}-\d{2})\b/);
-  if (isoMatch) {
-    observed_at = isoMatch[1];
-  } else {
-    // Relative date phrases like "today", "yesterday", "last week"
-    const relMatch = body.match(/\b(today|yesterday|last\s+week|last\s+month)\b/i);
-    if (relMatch) {
-      const phrase = relMatch[1].toLowerCase();
-      const now = new Date();
-      if (phrase === "today") {
-        observed_at = now.toISOString().slice(0, 10);
-      } else if (phrase === "yesterday") {
-        const d = new Date(now);
-        d.setDate(d.getDate() - 1);
-        observed_at = d.toISOString().slice(0, 10);
-      } else if (phrase.startsWith("last week")) {
-        const d = new Date(now);
-        d.setDate(d.getDate() - 7);
-        observed_at = d.toISOString().slice(0, 10);
-      } else if (phrase.startsWith("last month")) {
-        const d = new Date(now);
-        d.setMonth(d.getMonth() - 1);
-        observed_at = d.toISOString().slice(0, 10);
-      }
-    }
-  }
-
-  return { tags, source, observed_at, subjective };
-}
-
-/**
- * Attempt LLM enrichment of memory metadata.
- * Returns merged metadata fields on success.
- * On timeout, unreachable, or invalid JSON — returns empty result and emits a warning.
- * Never throws; always resolves.
- */
-async function runLlmEnrich(body: string): Promise<{
-  tags: string[];
-  description?: string;
-  observed_at?: string;
-}> {
-  const config = loadConfig();
-  if (!config.llm) {
-    warn("Warning: --enrich requires an LLM to be configured. Run `akm config set llm` to configure one.");
-    return { tags: [] };
-  }
-
-  const { chatCompletion, parseJsonResponse } = await import("./llm.js");
-
-  const prompt = `You are a memory tagger for a developer knowledge base.
-Given the memory text below, return ONLY a JSON object with these fields:
-- "tags": array of 1-5 short lowercase keyword tags
-- "description": one-sentence summary (optional)
-- "observed_at": ISO date (YYYY-MM-DD) if the text references a specific date (optional)
-
-Memory text:
-${body.slice(0, 2000)}
-
-Return ONLY the JSON object, no prose, no markdown fences.`;
-
-  try {
-    const { fetchWithTimeout: _fetchWithTimeout } = await import("./common.js");
-    // Use chatCompletion with a 10s timeout via AbortSignal in the underlying fetch.
-    // We wrap the entire call in a race with a 10s timeout promise.
-    const timeoutMs = 10_000;
-    const result = await Promise.race([
-      chatCompletion(
-        config.llm,
-        [
-          { role: "system", content: "Return only valid JSON. No prose." },
-          { role: "user", content: prompt },
-        ],
-        { maxTokens: 256, temperature: 0.1 },
-      ),
-      new Promise<never>((_, reject) => setTimeout(() => reject(new Error("LLM enrichment timed out")), timeoutMs)),
-    ]);
-
-    const parsed = parseJsonResponse<Record<string, unknown>>(result);
-    if (!parsed) {
-      warn("Warning: --enrich received invalid JSON from the LLM. Writing memory without enrichment.");
-      return { tags: [] };
-    }
-
-    const tags = Array.isArray(parsed.tags)
-      ? parsed.tags.filter((t): t is string => typeof t === "string" && t.trim().length > 0)
-      : [];
-
-    const description =
-      typeof parsed.description === "string" && parsed.description.trim() ? parsed.description.trim() : undefined;
-
-    const observed_at =
-      typeof parsed.observed_at === "string" && /^\d{4}-\d{2}-\d{2}$/.test(parsed.observed_at.trim())
-        ? parsed.observed_at.trim()
-        : undefined;
-
-    return { tags, description, observed_at };
-  } catch (err) {
-    const msg = err instanceof Error ? err.message : String(err);
-    warn(`Warning: --enrich failed (${msg}). Writing memory without enrichment.`);
-    return { tags: [] };
-  }
-}
 
 const rememberCommand = defineCommand({
   meta: {
@@ -2393,295 +2210,3 @@ function loadHints(detail: "normal" | "full" = "normal"): string {
   // Fallback for compiled binary — inline content
   return fallback;
 }
-
-const EMBEDDED_HINTS = `# akm CLI
-
-You have access to a searchable library of scripts, skills, commands, agents, knowledge documents, workflows, wikis, and memories via \`akm\`. Search your sources first before writing something from scratch.
-
-## Quick Reference
-
-\`\`\`sh
-akm search "<query>"                          # Search all sources
-akm curate "<task>"                          # Curate the best matches for a task
-akm search "<query>" --type workflow          # Filter to workflow assets
-akm search "<query>" --source both            # Also search registries
-akm show <ref>                                # View asset details
-akm workflow next <ref>                       # Start or resume a workflow
-akm remember "Deployment needs VPN access"    # Record a memory in your stash
-akm import ./notes/release-checklist.md       # Import a knowledge doc into your stash
-akm wiki list                                 # List available wikis
-akm wiki ingest <name>                        # Print the ingest workflow for a wiki
-akm feedback <ref> --positive|--negative      # Record whether an asset helped
-akm add <ref>                                 # Add a source (npm, GitHub, git, local dir)
-akm clone <ref>                               # Copy an asset to the working stash (optional --dest arg to clone to specific location)
-akm save                                      # Commit (and push if writable remote) changes in the primary stash
-akm registry search "<query>"                 # Search all registries
-\`\`\`
-
-## Primary Asset Types
-
-| Type | What \`akm show\` returns |
-| --- | --- |
-| script | A \`run\` command you can execute directly |
-| skill | Instructions to follow (read the full content) |
-| command | A prompt template with placeholders to fill in |
-| agent | A system prompt with model and tool hints |
-| knowledge | A reference doc (use \`toc\` or \`section "..."\` to navigate) |
-| workflow | Parsed steps plus workflow-specific execution commands |
-| memory | Recalled context (read the content for background information) |
-| vault | Key names only; use vault commands to inspect or load values safely |
-| wiki | A page in a multi-wiki knowledge base. For any wiki task, start with \`akm wiki list\`, then \`akm wiki ingest <name>\` for the workflow. Run \`akm wiki -h\` for the full surface. |
-
-When an asset meaningfully helps or fails, record that with \`akm feedback\` so
-future search ranking can learn from real usage.
-
-Run \`akm -h\` for the full command reference.
-`;
-
-const EMBEDDED_HINTS_FULL = `# akm CLI — Full Reference
-
-You have access to a searchable library of scripts, skills, commands, agents, knowledge documents, workflows, wikis, and memories via \`akm\`. Search your sources first before writing something from scratch.
-
-## Search
-
-\`\`\`sh
-akm search "<query>"                          # Search all sources
-akm curate "<task>"                          # Curate the best matches for a task
-akm search "<query>" --type workflow          # Filter by asset type
-akm search "<query>" --source both            # Also search registries
-akm search "<query>" --source registry        # Search registries only
-akm search "<query>" --limit 10               # Limit results
-akm search "<query>" --detail full            # Include scores, paths, timing
-\`\`\`
-
-| Flag | Values | Default |
-| --- | --- | --- |
-| \`--type\` | \`skill\`, \`command\`, \`agent\`, \`knowledge\`, \`workflow\`, \`script\`, \`memory\`, \`vault\`, \`wiki\`, \`any\` | \`any\` |
-| \`--source\` | \`stash\`, \`registry\`, \`both\` | \`stash\` |
-| \`--limit\` | number | \`20\` |
-| \`--format\` | \`json\`, \`jsonl\`, \`text\`, \`yaml\` | \`json\` |
-| \`--detail\` | \`brief\`, \`normal\`, \`full\`, \`summary\`, \`agent\` | \`brief\` |
-| \`--for-agent\` | boolean (deprecated — use \`--detail agent\`) | \`false\` |
-
-## Curate
-
-Combine search + follow-up hints into a dense summary for a task or prompt.
-
-\`\`\`sh
-akm curate "plan a release"                   # Pick top matches across asset types
-akm curate "deploy a Bun app" --limit 3       # Keep the summary shorter
-akm curate "review architecture" --type workflow # Restrict to one asset type
-\`\`\`
-
-## Show
-
-Display an asset by ref. Knowledge assets support view modes as positional arguments.
-
-\`\`\`sh
-akm show script:deploy.sh                     # Show script (returns run command)
-akm show skill:code-review                    # Show skill (returns full content)
-akm show command:release                      # Show command (returns template)
-akm show agent:architect                      # Show agent (returns system prompt)
-akm show workflow:ship-release                # Show parsed workflow steps
-akm show knowledge:guide toc                  # Table of contents
-akm show knowledge:guide section "Auth"       # Specific section
-akm show knowledge:guide lines 10 30          # Line range
-akm show knowledge:my-doc                    # Show content (local or remote)
-\`\`\`
-
-| Type | Key fields returned |
-| --- | --- |
-| script | \`run\`, \`setup\`, \`cwd\` |
-| skill | \`content\` (full SKILL.md) |
-| command | \`template\`, \`description\`, \`parameters\` |
-| agent | \`prompt\`, \`description\`, \`modelHint\`, \`toolPolicy\` |
-| knowledge | \`content\` (with view modes: \`full\`, \`toc\`, \`frontmatter\`, \`section\`, \`lines\`) |
-| workflow | \`workflowTitle\`, \`workflowParameters\`, \`steps\` |
-| memory | \`content\` (recalled context) |
-| vault | \`keys\`, \`comments\` |
-| wiki | \`content\` (same view modes as knowledge). For any wiki task, run \`akm wiki list\` then \`akm wiki ingest <name>\` for the workflow. |
-
-## Capture Knowledge While You Work
-
-\`\`\`sh
-akm remember "Deployment needs VPN access"     # Record a memory in your stash
-akm remember --name release-retro < notes.md   # Save multiline memory from stdin
-akm import ./docs/auth-flow.md                 # Import a file as knowledge
-akm import - --name scratch-notes < notes.md   # Import stdin as a knowledge doc
-akm workflow create ship-release               # Create a workflow asset in the stash
-akm workflow next workflow:ship-release        # Start or resume the next workflow step
-akm feedback skill:code-review --positive      # Record that an asset helped
-akm feedback agent:reviewer --negative         # Record that an asset missed the mark
-\`\`\`
-
-Use \`akm feedback\` whenever an asset materially helps or fails so future search
-ranking can learn from actual usage.
-
-## Wikis
-
-Multi-wiki knowledge bases (Karpathy-style). A stash-owned wiki lives at
-\`<stashDir>/wikis/<name>/\`; external directories or repos can also be registered
-as first-class wikis. akm owns lifecycle + raw-slug + lint + index regeneration
-for stash-owned wikis; page edits use your native Read/Write/Edit tools.
-
-\`\`\`sh
-akm wiki list                                  # List wikis (name, pages, raws, last-modified)
-akm wiki create research                       # Scaffold a new wiki
-akm wiki register ics-docs ~/code/ics-documentation # Register an external wiki
-akm wiki show research                         # Path, description, counts, last 3 log entries
-akm wiki pages research                        # Page refs + descriptions (excludes schema/index/log/raw)
-akm wiki search research "attention"           # Scoped search (equivalent to --type wiki --wiki research)
-akm wiki stash research ./paper.md             # Copy source into raw/<slug>.md (never overwrites)
-echo "..." | akm wiki stash research -         # stdin form
-akm wiki lint research                         # Structural checks: orphans, broken xrefs, uncited raws, stale index
-akm wiki ingest research                       # Print the ingest workflow for this wiki (no action)
-akm wiki remove research --force               # Delete pages/schema/index/log; preserves raw/
-akm wiki remove research --force --with-sources # Full nuke, including raw/
-\`\`\`
-
-**For any wiki task, start with \`akm wiki list\`, then \`akm wiki ingest <name>\`
-to get the step-by-step workflow.** Wiki pages are also addressable as
-\`wiki:<name>/<page-path>\` and show up in stash-wide \`akm search\` as
-\`type: wiki\`. Files under \`raw/\` and the wiki root infrastructure files
-\`schema.md\`, \`index.md\`, and \`log.md\` are not indexed and do not appear in
-search results. No \`--llm\` anywhere — akm never reasons about page content.
-
-## Vaults
-
-Encrypted-at-rest key/value stores for secrets. Each vault is a \`.env\`-format
-file at \`<stashDir>/vaults/<name>.env\`.
-
-\`\`\`sh
-akm vault create prod                         # Create a new vault
-akm vault set prod DB_URL postgres://...      # Set a key (or KEY=VALUE combined form)
-akm vault set prod DB_URL=postgres://...      # Combined KEY=VALUE form also works
-akm vault unset prod DB_URL                   # Remove a key
-akm vault list vault:prod                     # List key names (no values)
-akm vault show vault:prod                     # Same as list (alias)
-akm vault load vault:prod                     # Print export statements to source
-\`\`\`
-
-## Workflows
-
-Step-based workflows stored as \`<stashDir>/workflows/<name>.md\`.
-
-\`\`\`sh
-akm workflow template                         # Print a starter workflow template
-akm workflow create ship-release             # Scaffold a new workflow asset
-akm workflow start workflow:ship-release     # Start a new run
-akm workflow next workflow:ship-release      # Advance to the next step (or auto-start)
-akm workflow complete <run-id>               # Mark a step complete and advance
-akm workflow status <run-id>                 # Show current run status
-akm workflow resume <run-id>                 # Resume a blocked or failed run
-akm workflow list                            # List all workflow runs
-\`\`\`
-
-## Clone
-
-Copy an asset to the working stash or a custom destination for editing.
-
-\`\`\`sh
-akm clone <ref>                               # Clone to working stash
-akm clone <ref> --name new-name               # Rename on clone
-akm clone <ref> --dest ./project/.claude       # Clone to custom location
-akm clone <ref> --force                       # Overwrite existing
-akm clone "npm:@scope/pkg//script:deploy.sh"  # Clone from remote package
-\`\`\`
-
-When \`--dest\` is provided, \`akm init\` is not required first.
-
-## Save
-
-Commit local changes in a git-backed stash. Behaviour adapts automatically:
-
-- **Not a git repo** — no-op (silent skip)
-- **Git repo, no remote** — stage and commit only (the default stash always falls here)
-- **Git repo, has remote, not writable** — stage and commit only
-- **Git repo, has remote, \`writable: true\`** — stage, commit, and push
-
-\`\`\`sh
-akm save                                      # Save primary stash (timestamp message)
-akm save -m "Add deploy skill"               # Save with explicit message
-akm save my-skills                            # Save a named writable git stash
-akm save my-skills -m "Update patterns"      # Save named stash with message
-\`\`\`
-
-The \`--writable\` flag on \`akm add\` opts a remote git stash into push-on-save:
-
-\`\`\`sh
-akm add git@github.com:org/skills.git --provider git --name my-skills --writable
-\`\`\`
-
-## Add & Manage Sources
-
-\`\`\`sh
-akm add <ref>                                 # Add a source
-akm add @scope/stash                            # From npm (managed)
-akm add owner/repo                            # From GitHub (managed)
-akm add ./path/to/local/stash                   # Local directory
-akm add git@github.com:org/repo.git --provider git --name my-skills --writable
-akm enable skills.sh                          # Enable the skills.sh registry
-akm disable skills.sh                         # Disable the skills.sh registry
-akm list                                      # List all sources
-akm list --kind managed                       # List managed sources only
-akm remove <target>                           # Remove by id, ref, path, or name
-akm update --all                              # Update all managed sources
-akm update <target> --force                   # Force re-download
-\`\`\`
-
-## Registries
-
-\`\`\`sh
-akm registry list                             # List configured registries
-akm registry add <url>                        # Add a registry
-akm registry add <url> --name my-team         # Add with label
-akm registry add <url> --provider skills-sh   # Specify provider type
-akm registry remove <url-or-name>             # Remove a registry
-akm registry search "<query>"                 # Search all registries
-akm registry search "<query>" --assets        # Include asset-level results
-akm registry build-index                      # Build ./index.json
-akm registry build-index --out dist/index.json # Build to a custom path
-\`\`\`
-
-## Configuration
-
-\`\`\`sh
-akm config list                               # Show current config
-akm config get <key>                          # Read a value
-akm config set <key> <value>                  # Set a value
-akm config unset <key>                        # Remove a key
-akm config path --all                         # Show all config paths
-\`\`\`
-
-## Other Commands
-
-\`\`\`sh
-akm init                                      # Initialize working stash
-akm index                                     # Rebuild search index
-akm index --full                              # Full reindex
-akm list                                      # List all sources
-akm upgrade                                   # Upgrade akm using its install method
-akm upgrade --check                           # Check for updates
-akm help migrate 0.6.0                        # Print migration notes for a release (or: latest)
-akm hints                                     # Print this reference
-akm completions                               # Print bash completion script
-akm completions --install                     # Install completions
-\`\`\`
-
-## Output Control
-
-All commands accept \`--format\` and \`--detail\` flags:
-
-- \`--format json\` (default) — structured JSON
-- \`--format jsonl\` — one JSON object per line (streaming-friendly)
-- \`--format text\` — human-readable plain text
-- \`--format yaml\` — YAML output
-- \`--detail brief\` (default) — compact output
-- \`--detail normal\` — adds tags, refs, origins
-- \`--detail full\` — includes scores, paths, timing, debug info
-- \`--detail summary\` — metadata only (no content/template/prompt), under 200 tokens
-- \`--detail agent\` — agent-optimized output: strips non-actionable fields
-- \`--for-agent\` — deprecated alias for \`--detail agent\`
-
-Run \`akm -h\` or \`akm <command> -h\` for per-command help.
-`;
